Add new EnableFIPS image configuration option (#6075)
* feat: add new EnableFIPS image configuration option

This commit introduces a new EnableFIPS configuration option to the
KernelCommandline image schema. EnableFIPS is a boolean option that
controls whether the image tools should configure the system with FIPS
enabled or not.

To enable FIPS on a system, 3 parts must be present:
1. The argument fips=1 must be passed to the kernel via the kernel
commandline during boot. To do this, updates are needed to the grub.cfg
to make sure the command line parameter is passed when the kernel is
loaded into memory.

2. The dracut-fips package must be installed on the system. The
dracut-fips package contains FIPS self-tests which will halt the system
if any FIPS-approved kernel crypto algorithm is not functioning as
expected.

3. A boot= argument must be passed to the kernel via the kernel command
line during boot ONLY if the partition where /boot is mounted is
separate from the rootfs. This argument is needed for dracut-fips to
locate the kernel's .hmac file, which our kernel package places this in
/boot next to the kernel binary.

* docs: Add EnableFIPS documentation

* docs: add comments for bootPrefix and readOnlyRoot

Signed-off-by: Chris Co <chrco@microsoft.com>