ice: Fix multiple VF reset paths for SRIOV LAG
Original content for patch 0306 only fixed a single code path for VF reset.
VF reser can follow two different code paths depending on what triggered
the VF reset.  The original patch was abandoned upstream and replaced with
a new patch that creates a helper function and uses that in the multiple
reset paths.

This patch is to replace the abandoned upstream patch with the new complete
one.

Signed-off-by: Dave Ertman <david.m.ertman@intel.com>